backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
pydantic[email]
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
bcrypt==4.0.1
python-multipart>=0.0.9
httpx>=0.27.0
openpyxl>=3.1.0
PyMuPDF>=1.24.0
email-validator>=2.0.0
